requirement.txt
fastapi
uvicorn
pydantic
motor
python-dotenv
pywinauto
python-jose[cryptography]
bcrypt==4.0.1
python-multipart
slowapi
httpx
